Refactor: 구매금액 클래스 구현
- 구매금액을 의미하는 클래스인 BuyCash를 구현.
- 해당클래스에서 구매금액 적합성 검증을 하고, 구매금액과 관련된 기능을 수행함.
- 기존의 금액 검증 기능을 옮겨옴.
